requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
discord-webhook==0.13.0
discord.py==1.7.1
dnspython==2.0.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
multidict==5.1.0
packaging==20.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
yarl==1.6.3
